VE: Merge identical sub-refs based in reflistItemId
Sub-refs with identical Wikitext should be rendered as reuses in
VE.  To achive this we implement a lookup map to see
if a sub-ref with the same details was already added to the
ouput and the use that in the code to render it as a reuse.

In VE we can assume, that sub-refs with identical Wikitext are
merged in Parsoid already and that the Parsoid output will
link the reuses in the HTML by pointing them to the same item
in the references list.  So when converting to data in VE we are
able to use the id of the reference list items to match reuses.

In VE we're adding another lookup based on the refListId to the
converter step.  The lookup can then be used then to retrive listKey
and index from sub-refs that were already converted before and have
the identical refListId.

The lookup is stored with the Converter so it's persistent when
called on each DOM node conversion.  To make sure that the state
of that property is reset if the Converter start parsing another
document, we're making sure that the lookup is reset in that case.

Note, that this is only adding the VE changes, proving that they
do not break anything yet when converting Parsoid output that does
not include reuses.

Bug: T406467
Change-Id: I6f8342d4ff42efabf8bf9d6d16c63ecce69a37cc