Parsoid: Render sub-refs sharing the same details as reuse
Sub-refs with identical Wikitext should be rendered as reuses in
Parsoid.  To achive this we implement a lookup map to see if a
sub-ref with the same details was already added to the ouput.  We
can then reuse the reference to render it as a reuse.

Note, that we're not comparing if the rendered output is identical
but if the original Wikitext in the details attribute matches.

This will also make sure, that the nodes refer to the same reflist
item by using the reflist item's id.  This then can be used in VE
to match the nodes in there during conversion.

See I6f8342d4ff42efabf8bf9d6d16c63ecce69a37cc

Bug: T409065
Change-Id: I9ba1d3ec3a017b9cdfdc19f7c429b9f2bb41d65f