Move unrelated hook handlers out of ReferencePreviews class
Unfortunately we noticed to late that the patches I38c57d0 and
I5b3bf13 added unrelated code to the ReferencePreviews feature.
Neither the WikiEditor integration nor the VisualEditor integration
are related to ReferencePreviews. That's a pure read-mode feature.

We also lost the test for the ReferencePreviews integration in
I38c57d0 (the relevant "expects" assertion was removed). This patch
restores the previous test.

Bug: T406323
Change-Id: I56b3820dd93c55491f8bbb24123d498c4a5448c5