Replace unused services with (no-op) mocks in tests
Why bother interacting with a real service when the test in question
doesn't even interact with it?

The ReferencePreviewsGadgetsIntegration mocks are a little special.
A default mock from the `createMock()` method returns null (i.e. a
falsy value) everywhere. This is actually what we want in all these
cases. The mock acts as if the user doesn't have any gadget enabled.

Change-Id: I49b6f12b70efd6d360ad1934246f795d7dfaf2d1