VE: Don't create a synth ref when there's a LDR main ref
I improved the check that should make sure that the main content
is not lost when converting main refs to also include list defined
refs.  When we have a list defined ref that's providing we don't
have to add a synthetic one because currently there's no way to
remove these list defined refs so it's safe to assume that the
main content does not to be safed with synthetic alternative.

This seems also to fix the bug about empty lines when converting
a model that includes a LDR and a synthetic main ref for the same
content.

Bug: T411245
Change-Id: I99cafcc47bdfa016266c593d02e80249e862e102
(cherry picked from commit 529222067c8ded28989fce3cece594685dd25481)