Highlight ref search result when there is only one
We played around with a few hacks to make this widget behave more
like users expect. The problem is that the 2 visually distinct
elements (the input field and the list of search results) aren't
2 widgets. The tab key doesn't work as expected.

We partly fixed this in Icd1f7f5.

In addition, this patch highlights the 1st search result right away
when there is only 1 anyway. The user can see this and press enter.

I tried to apply this upstream in OOUI but it seems like this is
not possible.

Bug: T403588
Change-Id: I1a03c0df076b06459a7023881d636b8f3ac6572d