[cypress] skip tests when VisualEditor is not loaded
On CI, we would like to remove the recursive injection of extensions
dependencies. When Cite is a dependency, we are required to add
VisualEditor to have end-to-end tests to pass.

I have previously made `ext.cite.visualEditor` resource loader
registration to only happen when VisualEditor is loaded. Make the
integration tests to skip when that resource loader module is not
available.

Bug: T406323
Change-Id: I93f0a9a75e709eadd4251fe372487004f6820732
(cherry picked from commit 12e193ec67fb4be75447143b712aab2e73ebe1e4)