Only enable WikiEditor integration when it is loaded
On CI, we would like to remove the recursive injection of extensions
dependencies. When Cite is a dependency, we are required to add
WikiEditor to fullfil MediaWiki ResourcesTest::testValidDependencies
which otherwise would fail with:

  Dependencies that do not exist
  Failed asserting that two arrays are equal.
  --- Expected
  +++ Actual
  @@ @@
   Array (
       'ext.cite.wikiEditor' => Array (
  +        0 => 'ext.wikiEditor'
       )
   )

Move the `ext.cite.wikiEditor` registration from `extension.json` to the
PHP `ResourceLoaderRegisterModules` hook. Only load it when WikiEditor
is loaded.

Bug: T406323
Change-Id: I38c57d0cd68dd2469fdffcc585b0f06509530570