Stop relying on isSubRefWithMainBody flag
This is split from I49cc193 so we can merge it in two steps.
In this first patch all code stops using the previous
"isSubRefWithMainBody" flag but uses the "mainBody" id instead.

Note that the two flags have always been set together. Even if
stuff might come from an outdated cache, the new "mainBody" flag
is already guaranteed to be there.

Change-Id: I884f9661fe32f36ff0fb13fc6ad28e7e2ba65170