Allow searching for [1] with square brackets
This affects the "reuse an existing reference" feature.

It's already possible to search for the number. Unfortunately the
search result is often very noisy because the same (usually small)
numbers appear in many other places as well. Make it possible to
search for the number with square brackets if the user wants to.

This is optional and doesn't influence any existing behavior.

Change-Id: I3b8fa9aba122917ed294c72cbf9b22facbb222ff