VE: Introduce MWEditReferenceNodeAction and wire use cases
We already introduced a use case for a generic action that's just
about editing an existing reference node.  In there we can hide
some of the logic that then opens the Reference- or CitationDialog.

For references not using the Citation tooling we can already pass
a variable that allows to specify which node should be loaded.
Now this can be a different node then what's actually selected.

There's still work left to allow the latter for the CitationDialog.
See ticket.

Bug: T413760
Bug: T406022
Change-Id: Iae2eda7069db12f483e37240709d6a3aa897abb4